+++ types/index.d.ts
@@ -12,4 +12,4 @@
 /// <reference path="./runtime-fs.d.ts" />
 /// <reference path="./runtime-net.d.ts" />
 /// <reference path="./runtime-http.d.ts" />
-/// <reference path="./runtime-parsers.d.ts" />
+/// <reference path="./runtime-serialization.d.ts" />